streamlit_postgres/requirements.txt
plotly==5.18.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy
streamlit-aggrid
plotly
